php artisan doctrine:migrations:migrate --write-sql throws ErrorException (#46)
[ErrorException]
  Invalid argument supplied for foreach()

This happens as a result of the return from $sqlWriter->write which returns the result from file_put_contents. file_put_contents returns the number of bytes that were written to the file, or FALSE on failure. writeNotes expects an array of statements which needs to be looked up separately.